--- src/frontend/helpers/index.js
@@ -1,1 +1,0 @@
-export { getLikesData, postLikeData, deleteLikesData } from "./like_Helper";
