+++ src/frontend/services/index.js
@@ -1,2 +1,3 @@
 export { loginHandler, signupHandler, logoutHandler } from "./userServices";
 export { getVideos, getCategories } from "./dataServices";
+export { getLikesData, postLikeData, deleteLikesData } from "./likeServices";
